Enable updating rubygems itself to a certain version with --system x.y.z
This was not possible before because #which_to_update was used, which doesn't work when specifying the highest current version on the servers, as it specifies "> requested version".
IMHO, #which_to_update is not really what is semantically required for the rubygems-update updating. So instead, copy the smallish parts actually used and hardcode the rest inside #update_rubygems.
Also, distinguish in the error message between really being on the highest version, and the requested version not having been found.