RubyForge ticket [#29074] Fix update to specifically latest version
The tests found a bug in the previous fix - for the "update to latest" part, one should use ">", not ">=", so that in this case the latest version is not found, stopping unnecessary install.
Also, reusing a local variable makes my brain hurt, so try to come up with new names when the object behind "version" changes.